Rename BitcoinUrlBuilder to BitcoinUriBuilder
